requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
mlflow>=2.9.0
dvc>=3.0.0
open_clip_torch>=2.24.0
transformers>=4.41.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.4.0
timm>=1.0.0
psutil>=5.9.0
GPUtil>=1.4.0
PyYAML>=6.0
pytest>=7.4.0